+++ calc/Abstractions/ICalcBase.cs
@@ -1,0 +1,8 @@
+﻿namespace calc
+{
+    public interface ICalcBase
+    {
+        bool IsOperandus(char c);
+        bool IsOperator(char c);
+    }
+}
