+++ calc/Abstractions/ICalculator.cs
@@ -1,0 +1,7 @@
+﻿namespace calc
+{
+    public interface ICalculator
+    {
+        string Calculate(string input);
+    }
+}
