+++ calc/Abstractions/IOperation.cs
