+++ calc/Abstractions/IRPN.cs
@@ -1,0 +1,7 @@
+﻿namespace calc
+{
+    public interface IRPN
+    {
+        string LengyelFormaKonvertalas(string input);
+    }
+}
