+++ calc/Implementations/Operations/Division.cs
@@ -6,18 +6,8 @@ namespace calc.Operations
     {
         public double Operate(double quotient, double divisor)
         {
-
-            //try
-            //{
                 if (divisor == 0) throw new DivideByZeroException();
                 return quotient / divisor;
-            //}
-            //catch
-            //{
-            //    Console.WriteLine("DIVISION BY ZERO");
-            //    return 0;
-            //}
-
         }
     }
 }
