+++ calc/Implementations/Operations/Multiplication.cs
