+++ calc/Implementations/Operations/Power.cs
