+++ calc/Implementations/RPN.cs
@@ -5,7 +5,7 @@ using System.Text;
 
 namespace calc
 {
-    public class RPN : IRPN
+    public class RPN : CalcBase, IRPN
     {
         public String LengyelFormaKonvertalas(String input)
         {
@@ -48,42 +48,9 @@ namespace calc
             return String.Join(' ', formula);
         }
 
-
-        public String Calculate(String input)
-        {
-            var stack = new Stack<double>();
-            var array = input.Replace(" ", string.Empty).ToCharArray();
-
-            for (int i = 0; i < array.Length; i++)
-            {
-                char x = array[i];
-                if (IsOperandus(x))
-                {
-                    stack.Push(int.Parse(x.ToString()));
-                }
-                if (IsOperator(x))
-                {
-                    var second = stack.Pop();
-                    var first = stack.Pop();
-                    var operationresult = DoAction(x, first, second);
-                    stack.Push(operationresult);
-                }
-            }
-
-            return stack.Pop().ToString();
-        }
-
-
         #region Helpers
-        public bool IsOperator(char c)
-        {
-            return (c == OPERATION.MINUS || c == OPERATION.PLUS || c == OPERATION.MULTIPLY || c == OPERATION.DIVITE);
-        }
-        public bool IsOperandus(char c)
-        {
-            return (c >= '0' && c <= '9' || c == '.');
-        }
-        public int Prior(char c)
+       
+        protected int Prior(char c)
         {
             switch (c)
             {
@@ -99,25 +66,6 @@ namespace calc
                     return 3;
                 case OPERATION.POWER:
                     return 4;
-                default:
-                    throw new ArgumentException("Rossz parameter");
-            }
-        }
-
-        public double DoAction(char c, double a, double b)
-        {
-            switch (c)
-            {
-                case OPERATION.PLUS:
-                    return new Addition().Operate(a, b);
-                case OPERATION.MINUS:
-                    return new Subtraction().Operate(a, b);
-                case OPERATION.MULTIPLY:
-                    return new Multiplication().Operate(a, b);
-                case OPERATION.DIVITE:
-                    return new Division().Operate(a, b);
-                case OPERATION.POWER:
-                    return new Power().Operate(a, b);
                 default:
                     throw new ArgumentException("Rossz parameter");
             }
