+++ calc/Program.cs
@@ -9,11 +9,12 @@ namespace calc
             String str = "5 + ( ( 1 + 2 ) *  4 ) - 3";
             Console.WriteLine(str);
             IRPN manager = new RPN();
+            ICalculator calculator = new Calculator();
 
             var LFK = manager.LengyelFormaKonvertalas(str);
             Console.WriteLine(LFK);
 
-            var result = manager.Calculate(LFK);
+            var result = calculator.Calculate(LFK);
             Console.WriteLine(result);
             Console.ReadKey();
         }
