+++ calc/Static/Operands.cs
